enables Edit and delete in MasjeedDetails Page
